Perform Data Cleaning
Encoded categorical data into numerical data. Created an array out of it, looked at the shape, and converted it back into a dataframe. Created the other matrices of features that contained numerical data and dropped Attrition because we’ve been trying to determine and correlate reasons why based on the other features. Concatenated the categorical data and the numerical data into one DataFrame representing the matrices of features. Preprocessed the features with MinMaxScaler and fit_transform from sci-kit learn to scale all values to range from 0 to 1. Created the dependent variable dataframe from Attrition.