Built and trained with Random Forest Classifier Model
Imported the RandomForestClassifier from scikitlearn.ensemble module. Instantiated object with the class. Applied the fit method with the features and dependent training vars. Made predictions with predict passing the test features. Used the confusion_matrix again passing the prediction and test and plotted with a seaborn heatmap. Ended with a classification report looking at KPIs.