Built and trained artificial neural network classifier model
Used tensorflow and keras API, and built the model sequentially. We used a rectified linear unit (ReLU) activation function. Took a look at the layers with model dataframe.summary(). To train the model it was compiled, using Adam as the optimizer, and binary crossentropy as the loss function. Applied fit method to the model, made predictions, looked at predictions for values > 0.5. Plot the epochs throughout the training to visualize the progress in training for loss and accuracy. Plot the confusion matrix and finally looked at the classification report. Built the foundation to test three different classifiers. Next steps is to evaluate the results further, fine tune parameters to try to increase precision recall and f1-score over all.