Exploratory visualization, K Means, Elbow Method
Performed exploratory data visualization. Checked for any missing data with heatmap and verifying with summing null values for each column. For each null value, they were replaced with mean values. Also checked for duplications. Looked at data columns and noticed 1 column to be of no use so it was dropped in memory. Coded for number of columns to use in a for loop for subplotting a Kernel Density Estimate and histogram plot. Noted groups / averages of features from customers for several of the columns. Looked at a correlation map once more. WOrked through some of the theory and intuition behind K-Means and the within cluster sum of squares parameter that K is dependent on. With this knowledge, the optimal number of K was determined with a loop of 20 iterations. The data frame was scaled and determined an optimal K value from the plotted list.