Applying K-Means Clustering for Market Segmentation and PCA to visualize
Applying K-Means Clustering for market segmentation and combining it with principal component analysis. Had to do an inverse transform normalized data with kmeans cluster center to transform it back. Reviewing the data, took the clusters and put them in the column of our dataframe so we can make associations with the rest of the columns to a cluster, segmenting customers to a cluster. This was done by creating a subplot and iterating across columns and clusters and matching them to corresponding plots. Since we had 17 + 1 columns now we reduced the features down to 2 features and plotted the hue based on the cluster to plot the PCA of 2 components representing the 8 clusters of different types of customers that can be read through the histograms. The plot that visualizes clusters from the combo of K-Means and PCA provides a nice overview of what the clusters look like relative to each other in one nice plot!